[Kernel] Phase 4 (mamba): move JIT mamba operators into kernels.ops.mamba (RFC #29630)
Moves transfer_mamba inkling_sconv from sglang.jit_kernel into sglang.kernels.ops.mamba (classified);
compat shims at old paths (removed Phase 5); csrc stays in jit_kernel.
